Rename editor components
