+++ marketplace/cloudinary-assets/src/components/Editor/CloudinaryField.tsx
@@ -2,7 +2,7 @@ import * as React from 'react';
 import { Button } from '@contentful/forma-36-react-components';
 import { FieldExtensionSDK } from 'contentful-ui-extensions-sdk';
 import { CloudinaryResource } from '../../cloudinaryInterfaces';
-import { SortableComponent } from '../sortable/sortable';
+import { SortableComponent } from './SortableComponent';
 import { ExtensionParameters } from '../AppConfig/parameters';
 
 interface Props {
