+++ marketplace/cloudinary-assets/src/components/Editor/CloudinaryThumbnail.tsx
